Order prices types changed to decimal

See merge request !774

+++ kadena2.0/Kadena2.0.WebAPI/Contracts/ITaxEstimationService.cs
@@ -5,6 +5,6 @@ namespace Kadena.WebAPI.Contracts
 {
     public interface ITaxEstimationService
     {
-        Task<double> EstimateTotalTax(DeliveryAddress deliveryAddress);
+        Task<decimal> EstimateTotalTax(DeliveryAddress deliveryAddress);
     }
 }
